# Octree finalization optimization