--- a/chunky/src/java/se/llbit/chunky/renderer/scene/OctreeFinalizer.java
+++ b/chunky/src/java/se/llbit/chunky/renderer/scene/OctreeFinalizer.java
@@ -47,11 +47,10 @@ public class OctreeFinalizer {
         int z = cz + cp.z * 16 - origin.z;
         for (int cx = 0; cx < 16; ++cx) {
           int x = cx + cp.x * 16 - origin.x;
-          processBlock(worldTree, waterTree, palette, x, cy, z);
-
-          // only call hideBlocks for blocks on the edges of the chunk
+          // process blocks that are at the edge of the chunk, the other should have be taken care of during th loading
           if(cy == yMin || cy == yMax-1 || cz == 0 || cz == 15 || cx == 0 || cx == 15) {
             hideBlocks(worldTree, palette, x, cy, z, yMin, yMax, origin);
+            processBlock(worldTree, waterTree, palette, x, cy, z);
           }
         }
       }