--- a/chunky/src/java/se/llbit/math/BigPackedOctree.java
+++ b/chunky/src/java/se/llbit/math/BigPackedOctree.java
@@ -449,7 +449,7 @@ public class BigPackedOctree implements Octree.OctreeImplementation {
 
   @Override
   public void endFinalization() {
-    // There is a bunch of WHATEVER nodes we should try to merge
+    // There is a bunch of ANY_TYPE nodes we should try to merge
     finalizationNode(0);
   }