Related to #633 
Finalizing the octree is slow, especially when dealing with very large scene, because for each block we need to take action depending on its neighboring blocks, this means walking down the octree up to 10 times for each block. Walking down the octree is cheap but not free so, given the number of blocks this ends up costing a lot of time.
The idea of this PR is to do some of the computation requiring neighboring blocks when loading a chunk, when accessing blocks is cheaper as they are directly stored in a flat array. Blocks that are on the edge of a chunk don't have all of their neighbors available so those are still processed in the finalization phase as before.
While implementing this i also added the "whatever" type i suggested on discord that helps merging nodes more efficiently and added hook for octree implementation to be notified of when finalization takes place.

I did some measurement to compare the loading time before and after this PR. I loaded a 4k chunks scene three times without closing chunky in-between
Results:
|     | Loading time | Finalizing time | Total time |
|------|---------------|--------------------|---------------|
| Before, run 1 | 35s | 43s | 78s |
| Before, run 2 | 37s | 39s | 76s |
| Before, run 3 | 38s | 39s | 77s |
| **Before, average** | **37s** | **40s** | **77s** |
| After, run 1 | 27s | 7s | 34s |
| After, run 2 | 75s | 9s | 84s |
| After, run 3 | 39s | 9s | 48s |
| **After, average** | **46s** | **8s** | **55s** |

Results to take with a grain of salt, the variance for the after is huge (during the second run my computer froze for a bit, most likely due to swapping).

I would appreciate if someone with more memory could make some bigger scale tests