@aTom3333 2020-07-18
I need to do some measure to be sure but the introduction of the "whatever" type, although it does lower the node count (related to memory usage of the octree) in the end (after finalization), the peak node count seems to be higher and i'm not very pleased with that (if you didn't get it by now i like low memory usage :p)